+++ HISTORY.md
@@ -1,3 +1,9 @@
+0.1.19 / 20.12.2016
+===================
+
+  * (breaking change) Renamed `.rrui__select__option--selected` CSS class to `.rrui__select__option--focused`
+  * `<Select/>` now automatically scrolls the options list down to the selected item and also while navigating with arrow keys
+
 0.1.0 / 05.12.2016
 ==================